+++ testsuite/ada_lsp/completion_invisible_partially_qualified_prefix/p1-p2-subp.adb
@@ -1,0 +1,4 @@
+function P1.P2.Subp return Integer is
+begin
+   return 0;
+end P1.P2.Subp;