+++ testsuite/ada_lsp/completion_invisible_partially_qualified_prefix/p1-p3.ads
@@ -1,0 +1,3 @@
+package P1.P3 is
+   Obj : Integer := P2.
+end P1.P3;